apps/dash-dino/requirements.txt
torch
dash-labs==0.1.0
dash-bootstrap-components
spectra
colormath
requests
tinycss2
pandas
torchvision
Pillow
Flask-Caching